Make Construct much easier to modify on-the-fly.
Prior to this commit, modifying constructs was a delicate process
because you had to index relative to the first base in the construct.
So when you inserted or removed a base, you messed up all the indices
past that base.  This commit refactors Construct into a collection of
Domain objects, which each represent a mutable bit of sequence.  You can
insert or remove bases from a domain and only the indices in that domain
will be  changed.

The motivation for this refactor was to make it possible to write a
Monte Carlo search for designs.  A couple features were added with this
explicitly in mind.  First, constructs make it very easy to sample
aptamer insertion sites by allowing sequences to be "attached" to
domains.  Second, constructs can compare old folds to new folds, even
accounting for insertions and deletions.