Add loading dye back into the digestion master mix.
This was in a previous incarnation of the protocol, but I removed it
because the robot did a really poor job multidispensing ~50% glycerol
mixture.  But today, I realized that pipetting is much more accurate.

I also have realized that manually mixing reactions in 96-well PCR
plates is nearly impossible.  So I want the robot to do my mixing for
me, and this allows it to do that for the loading dye.