Disambiguate "spacer" in sgrna_helper.py
Previously I was using the term "spacer" to mean two different things:

1. The sequence in between the two halves of the aptamer.
2. The sequence targeted by Cas9.

The second usage doesn't really make sense (it's a very historical term)
but it is a standard term in the field.  Therefore I decided to refer to
the first usage as "splitter" instead.  This commit implements this
modest refactoring.