Add much more detailed transformation instructions.
As a consequence of all the optimization I did to get the MG1655
transformation to work well again, I realized that there were a lot of
parameters worth mentioning here.

I also added the MG1655 transformation to this protocol, because it's
more reliable to do the transformation in advance and to store the
libraries as glycerol stocks than it is to do the transformation the day
before and risk getting a bad yield and having to cancel FACS time.