Redesign the mh libraries.
I ran design simulations in the context of a number of different
spacers.  Compared to the simulations without spacers, these made the
nucleotides in the nexus and the ruler look more important and the
nucleotides in the hairpin less important.  I decided to put more faith
in the simulations with spacers.